+++ src/main/java/it/objectmethod/ecommerce/service/dto/CartDTO.java
@@ -4,23 +4,23 @@ import java.util.List;
 
 public class CartDTO {
 
-	private Integer id;
-	private Integer userId;
+	private Long id;
+	private Long userId;
 	private List<CartArticleDTO> cartArticles;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
